Move ActiveEffectConfig to the document sheets
