+++ src/foundry/foundry.js/applications/formApplications/documentSheets/activeEffectConfig.d.ts
@@ -1,4 +1,4 @@
-import { ConfiguredDocumentClass } from '../../../../types/helperTypes';
+import { ConfiguredDocumentClass } from '../../../../../types/helperTypes';
 
 declare global {
   /**
