+++ src/foundry/foundry.js/applications/formApplications/documentSheets/index.d.ts
@@ -1,3 +1,4 @@
+import './activeEffectConfig';
 import './actorSheet';
 import './itemSheet';
 import './journalSheet';
